Refactor: Eliminate service layer by integrating QuizAttemptService and QuestionService into QuizSession
- Deleted QuizAttemptService.java and QuestionService.java
- Enhanced QuizSession to handle quiz attempt lifecycle, answer submission, and completion(This was neccessary since these service classes where meant to enhance quizSession but instead they dublicated some code/added unneeded functions. for example, both classes had submitAnswer method and both did different things, which was unnacceptable
- Integrated question loading directly into QuizSession constructor
- Added database persistence methods to QuizSession for answer submission and scoring
- Consolidated test classes by merging functionality into QuizSessionTest
- Removed unnecessary service layer abstraction while maintaining all functionality
- Improved OOP design by giving QuizSession full responsibility for quiz session management
